입력 폼 스타일, 컴포넌트 분리하여 사용
